requirement.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
cfgv==3.3.1
distlib==0.3.2
Django==3.2.7
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
identify==2.2.14
nodeenv==1.6.0
numpy==1.21.2
openpyxl==3.0.8
pandas==1.3.3
platformdirs==2.3.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.7.2
